(#698) The best I could do without patching
Basically, what I did was making stuff work using the code that's already there, in terms of making the Emerald flames work, and activating the EmeraldBuff function.

The main problem is that there are moments where it does not work and or a) I find a way to make it work piggybacking on functions that AbstractRoom has (possible)

OR

I start patching. A problem for later.